split charge ingestion test into 2 files based on d08 or d18
