+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/DomainTests/ChargeIngestionTestsChargeInformation.cs
@@ -1,0 +1,117 @@
+﻿using System;
+using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
+using System.Xml.Linq;
+using Energinet.DataHub.Core.FunctionApp.TestCommon;
+using FluentAssertions;
+using FluentAssertions.Execution;
+using GreenEnergyHub.Charges.IntegrationTest.Core.Fixtures.FunctionApp;
+using GreenEnergyHub.Charges.IntegrationTest.Core.TestFiles.Charges;
+using GreenEnergyHub.Charges.IntegrationTest.Core.TestHelpers;
+using GreenEnergyHub.Charges.IntegrationTests.Fixtures;
+using Xunit;
+using Xunit.Abstractions;
+using Xunit.Categories;
+
+namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests
+{
+    [IntegrationTest]
+    public class ChargeIngestionTestChargeInformation
+    {
+        private const string EndpointUrl = "api/ChargeIngestion";
+        private const int SecondsToWaitForIntegrationEvents = 15;
+
+        [Collection(nameof(ChargesFunctionAppCollectionFixture))]
+        public class RunAsync : FunctionAppTestBase<ChargesFunctionAppFixture>, IAsyncLifetime
+        {
+            public RunAsync(ChargesFunctionAppFixture fixture, ITestOutputHelper testOutputHelper)
+                : base(fixture, testOutputHelper)
+            {
+            }
+
+            public Task InitializeAsync()
+            {
+                return Task.CompletedTask;
+            }
+
+            public Task DisposeAsync()
+            {
+                Fixture.ChargeCreatedListener.Reset();
+                Fixture.ChargePricesUpdatedListener.Reset();
+                Fixture.MessageHubMock.Clear();
+
+                // We need to clear host log after each test is done to ensure that we can assert on function executed
+                // on each test run because we only check on function name.
+                Fixture.HostManager.ClearHostLog();
+                return Task.CompletedTask;
+            }
+
+            [Fact]
+            public async Task When_ChargeInformationIsReceived_Then_ChargeCreatedIntegrationEventIsPublished()
+            {
+                // Arrange
+                var (request, correlationId) = Fixture.AsSystemOperator.PrepareHttpPostRequestWithAuthorization(
+                    EndpointUrl, ChargeDocument.TariffBundleWithCreateAndUpdate);
+                using var eventualChargeCreatedEvent = await Fixture
+                    .ChargeCreatedListener
+                    .ListenForMessageAsync(correlationId)
+                    .ConfigureAwait(false);
+
+                // Act
+                await Fixture.HostManager.HttpClient.SendAsync(request);
+
+                // Assert
+                var isChargeCreatedReceived = eventualChargeCreatedEvent
+                    .MessageAwaiter!
+                    .Wait(TimeSpan.FromSeconds(SecondsToWaitForIntegrationEvents));
+                isChargeCreatedReceived.Should().BeTrue();
+            }
+
+            [Fact]
+            public async Task Given_NewMessage_When_SenderIdDoesNotMatchAuthenticatedId_Then_ShouldReturnErrorMessage()
+            {
+                // Arrange
+                var (request, _) = Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
+                    EndpointUrl, ChargeDocument.ChargeDocumentWhereSenderIdDoNotMatchAuthorizedActorId);
+
+                // Act
+                var actual = await Fixture.HostManager.HttpClient.SendAsync(request);
+
+                // Assert
+                actual.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+                var errorMessage = await actual.Content.ReadAsStreamAsync();
+                var document = await XDocument.LoadAsync(errorMessage, LoadOptions.None, CancellationToken.None);
+                document.Element("Message")?.Value.Should().Be(ErrorMessageConstants.ActorIsNotWhoTheyClaimToBeErrorMessage);
+            }
+
+            [Fact(Skip = "Used for debugging ChargeCommandReceivedEventHandler")]
+            public async Task Given_UpdateRequest_When_GridAccessProviderPeeks_Then_MessageHubReceivesReply()
+            {
+                // Arrange - Create
+                var (createReq, correlationId) = Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
+                    EndpointUrl, ChargeDocument.CreateTariff);
+
+                var response = await Fixture.HostManager.HttpClient.SendAsync(createReq);
+                response.StatusCode.Should().Be(HttpStatusCode.Accepted);
+                await Fixture.MessageHubMock.AssertPeekReceivesRepliesAsync(correlationId);
+
+                // Arrange - Update
+                var (updateReq, updateCorrelationId) =
+                    Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
+                        EndpointUrl, ChargeDocument.UpdateTariff);
+
+                // Act
+                var actual = await Fixture.HostManager.HttpClient.SendAsync(updateReq);
+
+                // Assert
+                actual.StatusCode.Should().Be(HttpStatusCode.Accepted);
+
+                using var assertionScope = new AssertionScope();
+                // * Expect one for the confirmation (update)
+                var peekResults = await Fixture.MessageHubMock.AssertPeekReceivesRepliesAsync(updateCorrelationId);
+                peekResults.Should().ContainMatch("*ConfirmRequestChangeOfPriceList_MarketDocument*");
+            }
+        }
+    }
+}
