+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/DomainTests/ChargeIngestionTestsChargePrices.cs
@@ -41,7 +41,7 @@ using Xunit.Categories;
 namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests
 {
     [IntegrationTest]
-    public class ChargeIngestionTests
+    public class ChargeIngestionTestsChargePrices
     {
         private const string EndpointUrl = "api/ChargeIngestion";
         private const int SecondsToWaitForIntegrationEvents = 15;
@@ -83,23 +83,6 @@ namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests
             }
 
             [Fact]
-            public async Task Given_NewMessage_When_SenderIdDoesNotMatchAuthenticatedId_Then_ShouldReturnErrorMessage()
-            {
-                // Arrange
-                var (request, _) = Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
-                    EndpointUrl, ChargeDocument.ChargeDocumentWhereSenderIdDoNotMatchAuthorizedActorId);
-
-                // Act
-                var actual = await Fixture.HostManager.HttpClient.SendAsync(request);
-
-                // Assert
-                actual.StatusCode.Should().Be(HttpStatusCode.BadRequest);
-                var errorMessage = await actual.Content.ReadAsStreamAsync();
-                var document = await XDocument.LoadAsync(errorMessage, LoadOptions.None, CancellationToken.None);
-                document.Element("Message")?.Value.Should().Be(ErrorMessageConstants.ActorIsNotWhoTheyClaimToBeErrorMessage);
-            }
-
-            [Fact]
             public async Task When_ChargeIsReceived_Then_AHttp202ResponseWithEmptyBodyIsReturned()
             {
                 var request = Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
@@ -110,27 +93,6 @@ namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests
 
                 actualResponse.StatusCode.Should().Be(HttpStatusCode.Accepted);
                 responseBody.Should().BeEmpty();
-            }
-
-            [Fact]
-            public async Task When_ChargeInformationIsReceived_Then_ChargeCreatedIntegrationEventIsPublished()
-            {
-                // Arrange
-                var (request, correlationId) = Fixture.AsSystemOperator.PrepareHttpPostRequestWithAuthorization(
-                    EndpointUrl, ChargeDocument.TariffBundleWithCreateAndUpdate);
-                using var eventualChargeCreatedEvent = await Fixture
-                    .ChargeCreatedListener
-                    .ListenForMessageAsync(correlationId)
-                    .ConfigureAwait(false);
-
-                // Act
-                await Fixture.HostManager.HttpClient.SendAsync(request);
-
-                // Assert
-                var isChargeCreatedReceived = eventualChargeCreatedEvent
-                    .MessageAwaiter!
-                    .Wait(TimeSpan.FromSeconds(SecondsToWaitForIntegrationEvents));
-                isChargeCreatedReceived.Should().BeTrue();
             }
 
             [Fact]
@@ -380,34 +342,6 @@ namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests
                 notification.Should().NotContain("<cim:VATPayer>");
                 notification.Should().NotContain("<cim:transparentInvoicing>");
                 notification.Should().NotContain("<cim:taxIndicator>");
-            }
-
-            [Fact(Skip = "Used for debugging ChargeCommandReceivedEventHandler")]
-            public async Task Given_UpdateRequest_When_GridAccessProviderPeeks_Then_MessageHubReceivesReply()
-            {
-                // Arrange - Create
-                var (createReq, correlationId) = Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
-                    EndpointUrl, ChargeDocument.CreateTariff);
-
-                var response = await Fixture.HostManager.HttpClient.SendAsync(createReq);
-                response.StatusCode.Should().Be(HttpStatusCode.Accepted);
-                await Fixture.MessageHubMock.AssertPeekReceivesRepliesAsync(correlationId);
-
-                // Arrange - Update
-                var (updateReq, updateCorrelationId) =
-                    Fixture.AsGridAccessProvider.PrepareHttpPostRequestWithAuthorization(
-                    EndpointUrl, ChargeDocument.UpdateTariff);
-
-                // Act
-                var actual = await Fixture.HostManager.HttpClient.SendAsync(updateReq);
-
-                // Assert
-                actual.StatusCode.Should().Be(HttpStatusCode.Accepted);
-
-                using var assertionScope = new AssertionScope();
-                // * Expect one for the confirmation (update)
-                var peekResults = await Fixture.MessageHubMock.AssertPeekReceivesRepliesAsync(updateCorrelationId);
-                peekResults.Should().ContainMatch("*ConfirmRequestChangeOfPriceList_MarketDocument*");
             }
 
             [Fact]
